FEATURE-010: restructure folder and add start of summer, timeing, freq, dur calc
